+++ src/repl-window/repl-window-doc.ts
@@ -612,6 +612,9 @@ export function appendPrompt(onAppended?: OnAppendedCallback) {
   const prompt = getPrompt();
   if (!lastAppended.trimEnd().endsWith(prompt.trimEnd())) {
     appendLine(getPrompt(), onAppended);
+  } else if (onAppended) {
+    // Resolve promise though no append is actually needed
+    onAppended(undefined);
   }
 }